[CHORE] useCase 로직 구현
