+++ data/src/main/java/com/nexters/knownknowns/data/repository/AuthRepository.kt
@@ -10,6 +10,6 @@ interface AuthRepository {
     suspend fun setUserId(id: Long)
     fun getDeviceToken(): Flow<String>
     suspend fun setDeviceToken(token: String)
-    suspend fun registerUser(request: RegisterRequest): Flow<RegisterResponse>
-    suspend fun loginUser(deviceToken: String): Flow<LoginResponse>
+    suspend fun registerUser(request: RegisterRequest): Result<RegisterResponse>
+    suspend fun loginUser(deviceToken: String): Result<LoginResponse>
 }
