+++ domain/build.gradle.kts
@@ -44,4 +44,7 @@ dependencies {
     implementation(libs.koin.annotations)
     implementation(libs.koin.core)
     ksp(libs.koin.ksp)
+
+    // log
+    implementation(libs.timber)
 }
